test-apps/image-processor/requirements.txt
boto3
Pillow
numpy
aws-lambda-powertools[tracer]
